Creacion del componente SearchBar con los componentes de filtro combinado
