+++ client/src/components/Home/Home.jsx
@@ -2,8 +2,22 @@ import CategoryButtons from "../categoryButtons.jsx/CategoryButtons"
 import { Carousel } from './../Carousel/Carousel';
 import { Link } from "react-router-dom";
 import Loader from "../Loader/Loader";
+import { useEffect } from "react";
+import {setFilter} from "../../reduxToolkit/actions/productAction"
+import { useDispatch } from "react-redux";
 
 const Home = ()=> {
+    const dispatch = useDispatch()
+
+    useEffect(()=>{
+        dispatch(setFilter({
+            name: '',
+            categoryId: '',
+            order: ['name', 'asc'],
+        }))
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [])
+
     return(
         <div className="Home">
             <Loader></Loader>
